Another fix.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7395 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith-float.cpp
@@ -816,8 +816,9 @@ float RawFloat32::op(double a)
 {   return (float)a;
 }
 
+// Doing the rounding in two steps here is bad.
 float RawFloat32::op(LFlt a)
-{   return (float)a.floatval();
+{   return (float)(double)a.floatval();
 }
 
 double RawFloat::op(LispObject a)
